Sanitize project code on FwData import path
The ProjectCode regex was tightened to lowercase-only to match the LexBox
server rule, but MiniLcmImport still passed the FwData project name (e.g.
"Sena3") straight through as the code, so importing any project whose name
has uppercase letters or other disallowed characters now throws.

Promote the SanitizeProjectCode helper (previously a private copy in
SyncFixture) onto CrdtProjectsService and use it on both production paths
that derive a code from a name (the import path and the project-sync CLI).
